+++ minify-html/src/tests/mod.rs
@@ -176,7 +176,7 @@ fn test_viewport_attr_minification() {
 fn test_style_attr_minification() {
   eval_with_css_min(
     br#"<div style="color: yellow;"></div>"#,
-    br#"<div style=color:yellow></div>"#,
+    br#"<div style=color:#ff0></div>"#,
   );
   // `style` attributes are removed if fully minified away.
   eval_with_css_min(br#"<div style="  /*  */   "></div>"#, br#"<div></div>"#);