+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # minify-html changelog
 
+## Pending
+
+- Use [lightningcss](https://github.com/parcel-bundler/lightningcss) instead of css-minify, which is better maintained.
+  - BREAKING: The `minify_css_level_*` Cfg options no longer apply and have been removed.
+
 ## 0.12.0
 
 - Change CLI name to `minhtml` as it's a more concise command name and allows for `cargo install minhtml`.